+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Studio Changelog
 
+## Unreleased
+#### Changes
+
+#### Issues Resolved
+
+
 ## Upcoming release
 #### Changes
 
@@ -12,11 +18,14 @@
 * [[@kollivier](https://github.com/kollivier)] Sort countries by localized country name.
 * [[@jayoshih](https://github.com/jayoshih)] Added banner to redirect users from contentworkshop.learningequality.org to studio.learningequality.org
 * [[@benjaoming](https://github.com/benjaoming)] Moved Read the Docs out of main repo
+* [[@jayoshih](https://github.com/jayoshih)]Allow users to access invitations when they have access to the channel
+
 
 
 #### Issues Resolved
 * [#1004](https://github.com/learningequality/studio/issues/1004)
-* [#976](https://github.com/learningequality/studio/issues/1004)
+* [#976](https://github.com/learningequality/studio/issues/976)
+* [#1094](https://github.com/learningequality/studio/issues/1094)
 
 
 ## 2019-02-06 Release